Add modal popup for mobile data guide
On narrow views, instead of sliding the data guide in from the side,
show a modal view.
